fix: move the logic from container layer to pages layer
